Scripts added

+++ install.sh
@@ -1,26 +1,29 @@
 #!/bin/sh
 
+COLOR='\033[01;31m'	# bold red
+RESET='\033[00;00m'	# normal white
+
 OMZ=$HOME/.oh-my-zsh
 
 DOTPATH=`dirname $0`
 
-if [ -d $OMZ ]; then
+if [ -d "$OMZ" ]; then
 
 	echo "oh-my-zsh found!";
 
-	if [ ! -d $OMZ/custom ]; then
+	if [ ! -d "$OMZ/custom" ]; then
 		echo "mkdir custom"
 		mkdir $OMZ/custom
 	fi
 
-	if [ ! -d $OMZ/custom/themes ]; then
+	if [ ! -d "$OMZ/custom/themes" ]; then
 		echo "mkdir custom/themes"
 		mkdir $OMZ/custom/themes
 	fi
 
 	cp $DOTPATH/zsh/jdonat.zsh-theme $OMZ/custom/themes
 
-	if [ ! -d $OMZ/custom/plugins ]; then
+	if [ ! -d "$OMZ/custom/plugins" ]; then
 		echo "mkdir custom/plugins"
 		mkdir $OMZ/custom/plugins
 	fi
@@ -33,16 +36,35 @@ if [ -d $OMZ ]; then
 	cp $DOTPATH/.aliases $HOME
 
 
-	if [ ! -d $HOME/bin ]; then
+	if [ ! -d "$HOME/bin" ]; then
 		mkdir $HOME/bin
 	fi
 
-	if [ ! -d $HOME/Scripts ]; then
+	if [ ! -d "$HOME/Scripts" ]; then
 		mkdir $HOME/Scripts
 	fi
 
-	echo "\nInstalling Open Directory Scanner...\n"
-	rm -rf $HOME/Scripts/opendir
-	git clone https://gist.github.com/3285130.git $HOME/Scripts/opendir
+	RUBYPATH=`which ruby`
+	if [ -x "$RUBYPATH" ]; then
+		
+		echo "\nInstalling Open Directory Scanner...\n"
+		rm -rf $HOME/Scripts/opendir
+		git clone https://gist.github.com/3285130.git $HOME/Scripts/opendir
+		
+	else
+		echo "$COLOR - Ruby is not installed $RESET"
+	fi
 
-fi;
+	PHPPATH=`which php`
+	if [ -x "$PHPPATH" ]; then
+		
+		cp $DOTPATH/Scripts/argsr $HOME/Scripts
+		
+	else
+		echo "$COLOR - PHP is not installed $RESET"
+	fi
+
+
+else
+	echo "$COLOR - oh-my-zsh is not installed $RESET"
+fi
